dir graph tools and rnaabstract served the same purpose.
